prsima mongodb and some details
Some thoughts:
1. Clarify what is the client and what is the server:
(1) If the component involves the use of hooks, cache, and session processing, it is likely that the client side.
(2) If the current component is client-side and needs to call data, in addition to usequery, you can try the swr library, which is easier to use.

2.When using rich text
<div className={styles.description} dangerouslySetInnerHTML={{ __html: blog?.desc }} />

3. Global Prisma
Can be found in Prisma Best Practices

4. Prisma Grammar
When additional return data is needed -> $transaction

5. Processing Date
date-fns library